requirements.txt
torch>=2.11.0
torchvision>=0.26.0
numpy>=2.4.0
pandas>=2.3.0
scikit-learn>=1.8.0
matplotlib>=3.10.0
seaborn>=0.13.0
mlflow>=3.11.0
tqdm>=4.67.0
gdown>=5.2.0
Pillow>=12.1.0
